Layers images output to 2k display

+++ src/App/sys_timer.c
@@ -132,7 +132,7 @@ void		SYSTIMER_SetCountDown(uint8_t timer, uint32_t msecs)
 
 
 
-void		BUZZ_TimerOn(uint16_t time)
+void		BUZZ_TimerOn(uint32_t time)
 {
 	if (cfgConfig.buzzer == 0)
 		return;
@@ -144,7 +144,7 @@ void		BUZZ_TimerOn(uint16_t time)
 
 
 
-void		UVLED_TimerOn(uint16_t time)
+void		UVLED_TimerOn(uint32_t time)
 {
 	uvled_timer = time;
 	UVLED_On();
@@ -161,7 +161,7 @@ uint32_t	UVLED_TimerState()
 
 
 
-void		UVPAUSE_TimerOn(uint16_t time)
+void		UVPAUSE_TimerOn(uint32_t time)
 {
 	uvpause_timer = time;
 }
@@ -177,7 +177,7 @@ uint32_t	UVPAUSE_TimerState()
 
 
 
-void		UVFAN_TimerOn(uint16_t time)
+void		UVFAN_TimerOn(uint32_t time)
 {
 	FAN_LED_On();
 	uvfan_timer = time;
